pipeline/requirements.txt
spotipy
nats-py
pandas
snowflake-connector-python
snowflake-connector-python[pandas]
yacron
supervisor
